+++ packages/elasticsearch-store/src/elasticsearch-client/method-helpers/count.ts
@@ -22,7 +22,6 @@ export interface CountResponse {
     count: number;
 }
 
-
 export function convertCountParams(
     params: Record<string, any>,
     distribution: ElasticsearchDistribution,
